remove LLMTestSampleService
